+++ R/jfa_methods.R
@@ -73,14 +73,15 @@ print.summary.jfaPrior <- function(x, ...) {
 # Expected sample errors:       ", paste0(x[["x"]] * 100, "%"),"
 # Likelihood:                   ", x[["likelihood"]],"
 # Specifics:                    ", switch(x[["method"]],
-                                          "none" = "None",
+                                          "none" = "Default",
+                                          "uniform" = "Uniform",
                                           "median" = paste0("p(\u0398 < ", x[["materiality"]], ") = p(\u0398 > ", x[["materiality"]], ") = 0.5"),
-                                          "hypotheses" = paste0("p(\u0398 < ", x[["materiality"]], ") = ", x[["p.hmin"]],"; p(\u0398 > ", x[["materiality"]],") = ", x[["p.hplus"]]),
+                                          "hyp" = paste0("p(\u0398 < ", x[["materiality"]], ") = ", x[["p.hmin"]],"; p(\u0398 > ", x[["materiality"]],") = ", x[["p.hplus"]]),
                                           "arm" = paste0("IR = ", x[["ir"]], "; ICR = ", x[["icr"]], "; DR = ", x[["dr"]]),
                                           "bram" = paste0("Mode = ", x[["mode.prior"]], "; Upper bound = ", x[["ub.prior"]]),
                                           "sample" = paste0("Earlier sample of ", x[["n.prior"]], " items with ", x[["x.prior"]], " errors"),
                                           "factor" = paste0("Earlier sample of ", x[["n.prior"]], " items with ", x[["x.prior"]], " errors weighted by ", x[["factor"]]),
-                                          "custom" = paste0("\u03B1 = ", x[["alpha"]], "; \u03B2 = ", x[["beta"]])), "
+                                          "param" = paste0("\u03B1 = ", x[["alpha"]], "; \u03B2 = ", x[["beta"]])), "
 # ------------------------------------------------------------
 # Output: 
 #
@@ -263,7 +264,7 @@ summary.jfaPrior <- function(object, digits = 3, ...) {
   } else if (object[["method"]] == "bram") {
     out[["mode.prior"]] <- round(object[["specifics"]]$mode, digits)
     out[["ub.prior"]] <- round(object[["specifics"]]$ub, digits)
-  } else if (object[["method"]] == "hypotheses") {
+  } else if (object[["method"]] == "hyp") {
     out[["materiality"]] <- round(object[["materiality"]], digits + 2)
     out[["p.hmin"]] <- round(object[["specifics"]]$p.hmin, digits)
     out[["p.hplus"]] <- round(object[["specifics"]]$p.hplus, digits)
@@ -272,7 +273,7 @@ summary.jfaPrior <- function(object, digits = 3, ...) {
     out[["x.prior"]] <- object[["specifics"]]$x
     if (object[["method"]] == "factor")
       out[["factor"]] <- round(object[["specifics"]]$factor, digits)
-  } else if (object[["method"]] == 'custom') {
+  } else if (object[["method"]] == 'param') {
     out[["alpha"]] <- round(object[["specifics"]]$alpha, digits)
     out[["beta"]] <- round(object[["specifics"]]$beta, digits)
   }