+++ NEWS.md
@@ -2,7 +2,8 @@
 
 **New features**
 
-- Added `method = 'custom'` to function `auditPrior()` which takes as input the raw `alpha` and `beta` parameters of the prior distribution.
+- Added `method = 'param'` to function `auditPrior()` which takes as input the raw `alpha` and `beta` parameters of the prior distribution.
+- Added `method = 'uniform'` to function `auditPrior()` which constructs a uniform prior distribution (only available for the `binomial` and `hypergeometric` likelihood.)
 - Added a new vignette that describes the sampling methodology implemented in `jfa`.
 
 **Major changes**
@@ -37,8 +38,10 @@
     - `csMu` -> `cs.mu` (in `evaluation()`)
     - `records` -> `rows` (in `selection()`)
     - `mus` -> `values` (in `selection()`)
+    - `hypotheses` -> `hyp` (in `auditPrior()`)
 - `poisson` is now the default likelihood for all functions since it is the most conservative.
-- `times` (former `counts`) must now be indicated as a column name in the `data` instead of a vector.
+- The default prior distributions used when `method = 'none'` or `prior = FALSE` are now set to priors (`gamma(1, 0)`, `beta(1,0)`, `beta-binomial(1, 0)`) that yield the same results (with respect to sample sizes and upper limits) as classical procedures.
+- The `times` (former `counts`) argument in `evaluation()` must now be indicated as a column name in the `data` instead of a vector.
 - `nPrior` and `kPrior` have been removed from the `planning()` and `evaluation()` functions. All prior distributions must now be specified using `prior = TRUE` (noninformative priors) or using a call to `auditPrior()`.
 - Removed the `auditBF()` function since its value is available through `print(evaluation(materiality = x, prior = auditPrior(method = 'median', materiality = x)))`